Remove wrongly added files :(
